Use 32-bit integers instead of 16-bit for CAN messages

+++ src/interface/voltron_can/include/voltron_can/CanFrame.hpp
@@ -20,7 +20,7 @@ namespace Voltron {
   namespace Can {
     class CanFrame {
     public:
-      typedef uint16_t identifier_t;
+      typedef uint32_t identifier_t;
       typedef uint64_t data_t;
 
       // Construct from raw data
